organized and updated the file locations
