+++ src/main/java/frc/robot/RobotContainer.java
@@ -22,21 +22,8 @@ import edu.wpi.first.wpilibj2.command.DeferredCommand;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.DeviceIDs.controllerIDs;
 import frc.robot.commands.AddVisionMeasurement;
-import frc.robot.commands.ClimbingL1;
-import frc.robot.commands.ClimbingL2_3;
 import frc.robot.commands.ResetPose;
-import frc.robot.commands.Shooting;
-import frc.robot.commands.states.EjectingHopper;
-import frc.robot.commands.states.Intaking;
-import frc.robot.commands.states.ReverseShooter;
-import frc.robot.commands.states.Unclimb;
-import frc.robot.commands.states.PrepShoots.PrepAnywhere;
-import frc.robot.commands.states.PrepShoots.PrepDepot;
-import frc.robot.commands.states.PrepShoots.PrepNeutralToAlliance;
-import frc.robot.commands.states.PrepShoots.PrepNonOutpost;
-import frc.robot.commands.states.PrepShoots.PrepOpponentToAlliance;
-import frc.robot.commands.states.PrepShoots.PrepOutpost;
-import frc.robot.commands.states.PrepShoots.PrepTrench;
+import frc.robot.commands.states.*;
 import frc.robot.constants.ChoreoTraj;
 import frc.robot.constants.ConstDrivetrain;
 import frc.robot.constants.ConstMotion;
