+++ src/main/java/frc/robot/commands/states/Shooting.java
@@ -6,28 +6,33 @@ package frc.robot.commands.states;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.RobotContainer;
-import frc.robot.constants.ConstMotion;
+import frc.robot.constants.ConstRotors;
 import frc.robot.subsystems.StateMachine;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class Unclimb extends Command {
-  /** Creates a new Unclimb. */
+public class Shooting extends Command {
 
-  public Unclimb() {
+  /** Creates a new Shooting. */
+  public Shooting() {
     // Use addRequirements() here to declare subsystem dependencies.
+
     addRequirements(RobotContainer.stateMachineInstance);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    RobotContainer.stateMachineInstance.setRobotState(StateMachine.RobotState.UNCLIMB_L1);
-    RobotContainer.motionInstance.setClimberPosition(ConstMotion.EXTEND_CLIMBER);
+    RobotContainer.stateMachineInstance.setRobotState(StateMachine.RobotState.SHOOTING);
+    RobotContainer.rotorsInstance.setSerializerRollersSpeed(ConstRotors.SERIALIZER_ROLLERS_SPEED);
+    RobotContainer.rotorsInstance.setSerializerVFunnelSpeed(ConstRotors.SERIALIZER_V_FUNNEL_SPEED);
+    RobotContainer.rotorsInstance.setShooterTransferSpeed(ConstRotors.SHOOTER_TRANSFER_SPEED);
+    RobotContainer.rotorsInstance.setIntakeRollersSpeed(ConstRotors.INTAKE_ROLLER_SPEED);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
+    RobotContainer.drivetrainInstance.xBrake();
   }
 
   // Called once the command ends or is interrupted.
@@ -38,6 +43,6 @@ public class Unclimb extends Command {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return RobotContainer.motionInstance.isClimberAtPosition(ConstMotion.CLIMBER_TOLERANCE);
+    return true;
   }
 }
