+++ src/main/java/frc/robot/subsystems/StateMachine.java
@@ -9,9 +9,7 @@ import edu.wpi.first.epilogue.NotLogged;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.commands.ClimbingL1;
-import frc.robot.commands.ClimbingL2_3;
-import frc.robot.commands.Shooting;
+import frc.robot.commands.states.Climbing.*;
 import frc.robot.commands.states.*;
 import frc.robot.commands.states.Climbing.PrepClimb;
 import frc.robot.commands.states.PrepShoots.*;
